+++ editors/vscode/src/test/integration/suite/index.ts
@@ -3,7 +3,7 @@ import Mocha from "mocha";
 import { glob } from "glob";
 
 export async function run(): Promise<void> {
-  const mocha = new Mocha({ ui: "tdd", color: true, timeout: 20000 });
+  const mocha = new Mocha({ ui: "tdd", color: true, timeout: 30000 });
   const testsRoot = path.resolve(__dirname, "..");
 
   const files = await glob("**/*.test.js", { cwd: testsRoot });